refactor(security): extract estimateRoomNameStrength to shared utility
The _estimateRoomNameStrength function was duplicated verbatim in both
isInsecureRoomName.native.ts and isInsecureRoomName.web.ts. Extract it
into a dedicated shared module so it lives in a single place.

CHANGES
-------
- react/features/base/util/roomNameStrength.ts (new file): exports
  estimateRoomNameStrength() — the lightweight length-plus-variety
  heuristic that replaced zxcvbn.
- isInsecureRoomName.native.ts: removed local _estimateRoomNameStrength,
  import estimateRoomNameStrength from './roomNameStrength'.
- isInsecureRoomName.web.ts: same removal and import.

No behaviour change; TypeScript compilation (tsc:native + tsc:web)
passes with exit code 0.
